Remove check if RSA N is prime
The rsainvalid module contains basic checks for defect RSA keys.
The RSA N value is supposed to be a composite number, we were
checking if it is a prime.
This check has not much value (such keys would simply not work), but
it is computationally expensive. badkeys is ~10x faster without it.

Closes: https://github.com/badkeys/badkeys/issues/16