Add noqa comments for ruff
ruff warns about unused function arguments, which is usually a good
thing. In this specific case, however, we have a generic function schema
for RSA tests that expects 2 arguments (n, e). Yet, most tests do not
need the e, but we need to pass it anyway for the ones that do.