Prepare potential deprecation of blocklist formats
This allows adding a "deprecated" field to the update json data.
This can either be "true" to indicate an obsolete format without
further info, or a string message that will be shown to the user.
Furthermore, fail gracefully if "blocklist_url" is missing to
prepare for a situation where we do not want to provide updates
for obsolete formats any more.