Distinguish between valid and likely corrupt keys in smallfactors
Small prime factors are often an indication of a corrupt key, as
bit errors in an RSA modulus will usually lead to a number with
multiple small prime factors.
However, occasionally, "true" RSA keys with small factors can
be found in the wild: https://crt.sh/?id=21372703349

By checking whether the calculated p/q values are primes, we
can distinguish between likely corrupt and valid, but insecure,
RSA keys with small prime factors. This commit adds a subtest
for that. Given that this is only executed for broken/vulnerbale
keys, it is okay to run the "expensive" primality test.

Note that this does not consider the possibility of multiprime
RSA keys with small factors, but those are not widely used.